add multiterm search setting
lucene syntax requires each required term to be preceded by the '+' symbol. for most multi word searches we want this by default, the setting on the ui is set to on but can be toggled off.

tidies up some of the text and views. searcher now returns scores for the full search results through the api.