ignore last data point if it's early in the month and toggle
adds a toggle for not showing the last data point in the series if the period the data is reported from is incomplete. this is toggled on by default for days of the month before the 10th.

move javascript for trends page to separate script file which is minified to reduce response size.